fix(sql): preserve database error details across all adapters
Database errors include additional properties beyond just `message`:
- PostgreSQL: code, detail, hint, severity
- SQLite/LibSQL: code, errno
- DuckDB: code, detail

This change:
- Adds shared formatDbError() helper in shared/utils.ts
- Updates all adapters (postgres, sqlite, duckdb, json) to use it
- Exports formatDbError for consumers
- Ensures error messages include all available properties
