+++ packages/sql/src/postgres.ts
@@ -1,32 +1,6 @@
 import { DatabaseError, loadEnvConfig } from '@happyvertical/utils';
 import { Pool } from 'pg';
 import { DatabaseSchemaManager } from './schema-manager';
-
-/**
- * Formats a PostgreSQL error to include all relevant details
- * PostgreSQL errors include additional properties (code, detail, hint, severity)
- * that are lost when only accessing error.message
- *
- * @param error - The caught error object
- * @returns Formatted error string with all available PostgreSQL error details
- */
-function formatPgError(error: unknown): string {
-  const pgError = error as {
-    message?: string;
-    code?: string;
-    detail?: string;
-    hint?: string;
-    severity?: string;
-  };
-  const parts: string[] = [];
-  if (pgError.message) parts.push(pgError.message);
-  if (pgError.code) parts.push(`code=${pgError.code}`);
-  if (pgError.detail) parts.push(`detail=${pgError.detail}`);
-  if (pgError.hint) parts.push(`hint=${pgError.hint}`);
-  if (pgError.severity) parts.push(`severity=${pgError.severity}`);
-  return parts.length > 0 ? parts.join(', ') : String(error);
-}
-
 import {
   generateAddColumnStatement,
   generateCreateIndexStatement,
@@ -45,7 +19,7 @@ import type {
   TableSchemaInfo,
   TransactionHandle,
 } from './shared/types';
-import { buildWhere } from './shared/utils';
+import { buildWhere, formatDbError } from './shared/utils';
 
 /**
  * Configuration options for PostgreSQL database connections
@@ -147,7 +121,7 @@ async function createTablesFromSchemas(
         }
       }
     } catch (error) {
-      const errMsg = formatPgError(error);
+      const errMsg = formatDbError(error);
       throw new DatabaseError(
         `Failed to create table ${tableName} from schema: ${errMsg}`,
         {
@@ -330,7 +304,7 @@ export async function getDatabase(
         table,
         sql: query,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -356,7 +330,7 @@ export async function getDatabase(
         table,
         sql: query,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -392,7 +366,7 @@ export async function getDatabase(
         table,
         sql,
         values: [...values, ...whereValues],
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -441,7 +415,7 @@ export async function getDatabase(
         sql,
         values,
         conflictColumns,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -509,7 +483,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to delete records from table', {
         table,
         where,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -554,7 +528,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to count records in table', {
         table,
         where,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -630,7 +604,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to execute pluck query', {
         sql,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -654,7 +628,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to execute single query', {
         sql,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -678,7 +652,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to execute many query', {
         sql,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -701,7 +675,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to execute query', {
         sql,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -727,7 +701,7 @@ export async function getDatabase(
       throw new DatabaseError('Failed to execute raw query', {
         sql,
         values,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -1439,7 +1413,7 @@ export async function getDatabase(
     } catch (e) {
       throw new DatabaseError('Failed to retrieve table schema', {
         table,
-        originalError: formatPgError(e),
+        originalError: formatDbError(e),
       });
     }
   };
@@ -1470,7 +1444,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to add column to table', {
           table,
           column: column.name,
-          originalError: formatPgError(e),
+          originalError: formatDbError(e),
         });
       }
     },
@@ -1498,7 +1472,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to create index on table', {
           table,
           index: index.name,
-          originalError: formatPgError(e),
+          originalError: formatDbError(e),
         });
       }
     },
