+++ packages/sql/src/sqlite.ts
@@ -19,7 +19,7 @@ import type {
   TableSchemaInfo,
   TransactionHandle,
 } from './shared/types';
-import { buildWhere } from './shared/utils';
+import { buildWhere, formatDbError } from './shared/utils';
 
 /**
  * Connection cache for in-memory databases with memoryId
@@ -324,7 +324,7 @@ export async function getDatabase(
           table,
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -353,7 +353,7 @@ export async function getDatabase(
           table,
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -380,7 +380,7 @@ export async function getDatabase(
           table,
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -421,7 +421,7 @@ export async function getDatabase(
           table,
           sql,
           values: [...values, ...whereValues],
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -477,7 +477,7 @@ export async function getDatabase(
           sql,
           values,
           conflictColumns,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -876,7 +876,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to execute pluck query', {
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -901,7 +901,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to execute single query', {
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -926,7 +926,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to execute many query', {
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -950,7 +950,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to execute query', {
           sql,
           values,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -987,7 +987,7 @@ export async function getDatabase(
         throw new DatabaseError('Failed to execute raw query', {
           sql,
           args,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -1135,7 +1135,7 @@ export async function getDatabase(
       } catch (e) {
         throw new DatabaseError('Failed to retrieve table schema', {
           table,
-          originalError: e instanceof Error ? e.message : String(e),
+          originalError: formatDbError(e),
         });
       }
     };
@@ -1166,7 +1166,7 @@ export async function getDatabase(
           throw new DatabaseError('Failed to add column to table', {
             table,
             column: column.name,
-            originalError: e instanceof Error ? e.message : String(e),
+            originalError: formatDbError(e),
           });
         }
       },
@@ -1197,7 +1197,7 @@ export async function getDatabase(
           throw new DatabaseError('Failed to create index on table', {
             table,
             index: index.name,
-            originalError: e instanceof Error ? e.message : String(e),
+            originalError: formatDbError(e),
           });
         }
       },
